Move rule stubs in own package under src/test/java
